+++ CHANGELOG.md
@@ -7,7 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 - Support concurrent hardware wallets logins
-
+- Liquid network: support for watch-only wallets
+- Watch-only wallets: support to delete credentials
+ 
 ### Changed
 - Improve performance of login with hardware wallets